Cleaning up game event API documentation.
