+++ Assets/Scripts/Game/Events/GameEventBinding.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 15152d22cf284de69f81f7953831486e
